+++ source/egov-erp/egov-egi/src/main/java/org/egov/infstr/workflow/admin/WorkflowAdminService.java
@@ -42,7 +42,7 @@ public class WorkflowAdminService {
         return daoService.findAllByNamedQuery(WorkflowTypes.TYPE_LIKE_NAME, name.toLowerCase(Locale.ENGLISH) + "%");
     }
 
-    public String reassignWFItem(final String wfType, final String stateId, final Integer newUserId) {
+    public String reassignWFItem(final String wfType, final String stateId, final Long newUserId) {
         String status = "ERROR";
         final List<StateAware> stateAwares = inboxRenderServiceDeligate.getWorkflowItems(wfType, stateId);
         if (!stateAwares.isEmpty()) {
